Move logic up one level to Quiz from IntervalExerciseController
