+++ src/IntervalExerciseController.js
@@ -1,53 +1,24 @@
 import React, { Component } from "react";
-import Teoria from "teoria";
-
+import PropTypes from "prop-types";
 import IntervalExercise from "./IntervalExercise.js";
 
-const POSSIBLE_INTERVALS = [
-    Teoria.interval('m2'),
-    Teoria.interval('M2'),
-    Teoria.interval('m3'),
-    Teoria.interval('M3'),
-    Teoria.interval('P4'),
-    Teoria.interval('d5'),
-    Teoria.interval('P5'),
-    Teoria.interval('m6'),
-    Teoria.interval('M6'),
-    Teoria.interval('m7'),
-    Teoria.interval('M7'),
-    // TODO: include octave
-
-]
-
-function getRandomInt(max) {
-  return Math.floor(Math.random() * Math.floor(max));
-}
 
 class IntervalExerciseController extends Component {
-    constructor() {
-        super()
-
-        this.state = {
-            interval: POSSIBLE_INTERVALS[getRandomInt(POSSIBLE_INTERVALS.length)],
-            submittedAnswers: []
-        }
-        this.handleAnswerClick = this.handleAnswerClick.bind(this)
-    }
-
-    handleAnswerClick(e) {
-        this.setState({
-            submittedAnswers: this.state.submittedAnswers.concat([e.currentTarget.value])
-        })
-    }
-
     render() {
         return <IntervalExercise
-            interval={this.state.interval}
-            possibleAnswers={POSSIBLE_INTERVALS}
-            submittedAnswers={this.state.submittedAnswers}
-            onAnswerClick={this.handleAnswerClick}
+            interval={this.props.interval}
+            possibleAnswers={this.props.possibleAnswers}
+            submittedAnswers={this.props.submittedAnswers}
+            onAnswerClick={this.props.onAnswerClick}
         />
     }
+}
+
+IntervalExerciseController.propTypes = {
+    interval: PropTypes.object, // Teoria interval
+    possibleAnswers: PropTypes.arrayOf(PropTypes.object), // Array of Teoria intervals
+    submittedAnswers: PropTypes.arrayOf(PropTypes.string), // Array of string ids  of Teoria intervals
+    onAnswerClick: PropTypes.func,
 }
 
 export default IntervalExerciseController
